Revert "feat: self hosted onboarding (#5057)"
This reverts commit 52cf21bec5ca9919dea3b1e84e4e210523a9a30f.
