+++ frontend/common/types/requests.ts
@@ -49,20 +49,6 @@ export type CreateVersionFeatureState = {
   sha: string
   featureState: FeatureState
 }
-
-export type LoginRequest = {
-  email: string
-  password: string
-}
-export type RegisterRequest = {
-  email: string
-  first_name: string
-  last_name: string
-  password: string
-  superuser?: boolean
-  organisation_name?: string
-  marketing_consent_given?: boolean
-}
 export type Req = {
   getSegments: PagedRequest<{
     q?: string
@@ -635,16 +621,5 @@ export type Req = {
   getSplitTest: PagedRequest<{
     conversion_event_type_id: string
   }>
-  createOnboarding: {
-    first_name: string
-    last_name: string
-    email: string
-    password: string
-    contact_consent_given: boolean
-    organisation_name: string
-    superuser: boolean
-  }
-  getBuildVersion: {}
-  createOnboardingSupportOptIn: {}
   // END OF TYPES
 }
