+++ frontend/common/types/responses.ts
@@ -691,27 +691,6 @@ export type HealthProvider = {
   webhook_url: number
 }
 
-export type Version = {
-  tag: string
-  backend_sha: string
-  frontend_sha: string
-  frontend: {
-    ci_commit_sha?: string
-    image_tag?: string
-  }
-  backend: {
-    ci_commit_sha: string
-    image_tag: string
-    has_email_provider: boolean
-    is_enterprise: boolean
-    is_saas: boolean
-    'self_hosted_data'?: {
-      'has_users': boolean
-      'has_logins': boolean
-    }
-  }
-}
-
 export type PConfidence =
   | 'VERY_LOW'
   | 'LOW'
@@ -755,10 +734,6 @@ export type Webhook = {
   enabled: boolean
   created_at: string
   updated_at: string
-}
-
-export type AccountModel = User & {
-  organisations: Organisation[]
 }
 
 export type IdentityTrait = {
@@ -903,6 +878,5 @@ export type Res = {
   identityTraits: IdentityTrait[]
   conversionEvents: PagedResponse<ConversionEvent>
   splitTest: PagedResponse<SplitTestResult>
-  onboardingSupportOptIn: { id: string }
   // END OF TYPES
 }
