--- frontend/common/utils/isFreeEmailDomain.ts
@@ -1,7 +1,0 @@
-import freeEmailDomains from 'free-email-domains'
-
-export default function (value: string | null | undefined) {
-  if (!value) return false
-  const domain = value?.split('@')?.[1]
-  return freeEmailDomains.includes(domain)
-}
