+++ frontend/global.d.ts
@@ -10,15 +10,9 @@ export type OpenConfirm = {
   noText?: string
 }
 import { TooltipProps } from './web/components/Tooltip'
-type CrispCommand = [command: string, ...args: any[]]
-type Crisp = {
-  // The push method accepts a CrispCommand array.
-  push: (command: CrispCommand) => void
-}
 
 export declare const openModal: (name?: string) => Promise<void>
 declare global {
-  const $crisp: Crisp
   const openModal: (
     title: ReactNode,
     body?: ReactNode,
@@ -51,7 +45,4 @@ declare global {
   const closeModal2: () => void
   const toast: (message: string) => void
   const Tooltip: FC<TooltipProps>
-  interface Window {
-    $crisp: Crisp
-  }
 }
