+++ frontend/web/components/App.js
@@ -13,6 +13,7 @@ import ButterBar from './ButterBar'
 import AccountSettingsPage from './pages/AccountSettingsPage'
 import Headway from './Headway'
 import ProjectStore from 'common/stores/project-store'
+import getBuildVersion from 'project/getBuildVersion'
 import { Provider } from 'react-redux'
 import { getStore } from 'common/store'
 import { resolveAuthFlow } from '@datadog/ui-extensions-sdk'
@@ -106,6 +107,7 @@ const App = class extends Component {
       })
       amplitude.add(sessionReplayTracking)
     }
+    getBuildVersion()
     this.state.projectId = this.getProjectId(this.props)
     if (this.state.projectId) {
       AppActions.getProject(this.state.projectId)
