+++ frontend/web/components/pages/HomeAside.tsx
@@ -28,7 +28,6 @@ import { components } from 'react-select'
 import SettingsIcon from 'components/svg/SettingsIcon'
 import BuildVersion from 'components/BuildVersion'
 import { useGetHealthEventsQuery } from 'common/services/useHealthEvents'
-import Resources from 'components/Resources'
 import Constants from 'common/constants'
 
 type HomeAsideType = {
@@ -388,13 +387,8 @@ const HomeAside: FC<HomeAsideType> = ({
                     clearableValue={false}
                   />
                 </div>
-                <div
-                  style={{ width: 260 }}
-                  className='text-muted position-fixed bottom-0 p-2 fs-caption d-flex flex-column gap-4'
-                >
-                  <Resources />
-                  <BuildVersion />
-                </div>
+
+                <BuildVersion />
               </div>
             )
           }}
