+++ frontend/web/project/api.js
@@ -1,7 +1,5 @@
 import * as amplitude from '@amplitude/analytics-browser'
 import data from 'common/data/base/_data'
-import isFreeEmailDomain from 'common/utils/isFreeEmailDomain'
-
 const enableDynatrace = !!window.enableDynatrace && typeof dtrum !== 'undefined'
 import { loadReoScript } from 'reodotdev'
 
@@ -68,7 +66,7 @@ global.API = {
     if (enableDynatrace && user?.id) {
       dtrum.identifyUser(`${user.id}`)
     }
-    Utils.setupCrisp()
+
     if (Project.heap) {
       heap.identify(id)
       const user = AccountStore.model
@@ -169,7 +167,7 @@ global.API = {
       .then(() => {
         const organisation = AccountStore.getOrganisation()
         const emailDomain = `${user?.email}`?.split('@')[1] || ''
-        const freeDomain = isFreeEmailDomain(emailDomain)
+        const freeDomain = freeEmailDomains.includes(emailDomain)
         if (
           !freeDomain &&
           typeof delighted !== 'undefined' &&
@@ -273,7 +271,10 @@ global.API = {
         if (plan && plan.includes('start-up')) {
           return planNames.startup
         }
-        if (global.flagsmithVersion?.backend.is_enterprise || (plan && plan.includes('enterprise'))) {
+        if (
+          global.flagsmithVersion?.backend.is_enterprise ||
+          (plan && plan.includes('enterprise'))
+        ) {
           return planNames.enterprise
         }
         return planNames.free
