+++ frontend/web/styles/_variables.scss
@@ -352,7 +352,6 @@ $scrollbar-thumb-dark-hover: $bg-dark300;
 
 $theme-colors: (
         light200: $bg-light200,
-        light300: $bg-light300,
         primary: $primary,
         primary900: $primary900,
         danger: $danger,
