+++ frontend/web/styles/project/_project-nav.scss
@@ -1,12 +1,4 @@
-.onboarding {
-  background-color: $bg-light200;
-  nav {
-    display: none;
-  }
-  .butter-bar {
-    display: none;
-  }
-}
+
 nav a {
   font-weight: 500;
   &.active {
