+++ frontend/web/styles/project/_spacing-utils.scss
@@ -430,6 +430,3 @@
 .mx-auto {
     margin-left: auto !important;
 }
-.row-gap-4 {
-    row-gap: $spacer;
-}
