+++ frontend/web/styles/project/_type.scss
@@ -140,6 +140,3 @@ a.dark {
     color: $primary;
   }
 }
-.letter-spacing {
-  letter-spacing: 1px;
-}
